requirements.txt
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
kafka-python==2.0.2
kombu==5.2.4
kombu-sqlalchemy==1.1.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pydantic==1.10.7
pytesseract==0.3.10
python-multipart==0.0.6
pytz==2023.3
redis==4.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
sqlalchemy-json==0.5.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
